barbuzz/backend/requirements.txt
django-allauth[mfa,socialaccount]>=65.7.0
qrcode >= 7.0.0
drf-spectacular == 0.28.0
djangorestframework >= 3.15.2,<4
django-ninja >= 1.3.0,<2
psycopg2-binary>=2.9.9
dj-rest-auth >= 7.0.1
